+++ Changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Master
+
+- Reorganize Hutch config - instead of setting config values when connecting to Hutch when using publisher, do it right after the initialization.
+
 ## 0.2.0
 - Implement "Safe Producer" extension allowing to retry delivery later in an automated way when publishing of the message fails.